Deprecate goal share-engine-core-classpath
