+++ src/main/java/ch/ivyteam/ivy/maven/ShareEngineCoreClasspathMojo.java
@@ -16,7 +16,12 @@
 
 package ch.ivyteam.ivy.maven;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
@@ -26,7 +31,7 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
 
-import ch.ivyteam.ivy.maven.engine.EngineClassLoaderFactory;
+import ch.ivyteam.ivy.maven.engine.EngineClassLoaderFactory.OsgiDir;
 import ch.ivyteam.ivy.maven.util.MavenProperties;
 
 /**
@@ -34,22 +39,58 @@ import ch.ivyteam.ivy.maven.util.MavenProperties;
  * <code>ivy.engine.core.classpath</code>.
  *
  * @since 6.2.0
+ * @deprecated Sharing the classpath of the Axon Ivy Engine is no longer needed, because all
+ *             dependencies are available as Maven dependencies.
  */
+@Deprecated(since = "14.0.0")
 @Mojo(name = ShareEngineCoreClasspathMojo.GOAL)
 public class ShareEngineCoreClasspathMojo extends AbstractEngineMojo {
 
+  private static final List<String> ENGINE_LIB_DIRECTORIES = Arrays.asList(
+      OsgiDir.INSTALL_AREA + "/" + OsgiDir.LIB_BOOT,
+      OsgiDir.PLUGINS,
+      OsgiDir.INSTALL_AREA + "/configuration/org.eclipse.osgi", // unpacked jars from OSGI bundles
+      "webapps" + File.separator + "ivy" + File.separator + "WEB-INF" + File.separator + "lib");
+
+  @Deprecated
   public static final String GOAL = "share-engine-core-classpath";
 
+  @Deprecated
   public static final String IVY_ENGINE_CORE_CLASSPATH_PROPERTY = "ivy.engine.core.classpath";
 
+  @Deprecated
   @Parameter(property = "project", required = true, readonly = true)
   MavenProject project;
 
+  @Deprecated
   @Override
   public void execute() throws MojoExecutionException, MojoFailureException {
-    List<Path> ivyEngineClassPathFiles = EngineClassLoaderFactory.getIvyEngineClassPathFiles(identifyAndGetEngineDirectory());
+    List<Path> ivyEngineClassPathFiles = getIvyEngineClassPathFiles(identifyAndGetEngineDirectory());
     String propertyValue = StringUtils.join(ivyEngineClassPathFiles, ",");
     MavenProperties properties = new MavenProperties(project, getLog());
     properties.setMavenProperty(IVY_ENGINE_CORE_CLASSPATH_PROPERTY, propertyValue);
+  }
+
+  private static List<Path> getIvyEngineClassPathFiles(Path engineDirectory) {
+    if (engineDirectory == null) {
+      return List.of();
+    }
+
+    var classPathFiles = new ArrayList<Path>();
+    for (String libDirPath : ENGINE_LIB_DIRECTORIES) {
+      var jarDir = engineDirectory.resolve(libDirPath);
+      if (!Files.isDirectory(jarDir)) {
+        continue;
+      }
+      try (var walker = Files.walk(jarDir)) {
+        var jars = walker
+            .filter(p -> p.getFileName().toString().endsWith(".jar"))
+            .toList();
+        classPathFiles.addAll(jars);
+      } catch (IOException ex) {
+        throw new RuntimeException(ex);
+      }
+    }
+    return classPathFiles;
   }
 }
